+++ src/semantic_tokens/kotlin.rs
@@ -195,7 +195,7 @@ fn kotlin_prop_token(node: Node<'_>, src: &Source<'_>, out: &mut Vec<RawToken>) 
         }
     } else if let Some(multi) = first_child_of_kind(node, KIND_MULTI_VAR_DECL) {
         for i in 0..multi.named_child_count() {
-            if let Some(vd) = multi.named_child(i) {
+            if let Some(vd) = multi.named_child(i as u32) {
                 if let Some(name) = child_ident(vd) {
                     push_token(name, token_type, mods, src, out);
                 }
